refactor: change all tests from post for update and delete to put and delete mapping

+++ src/main/java/com/pharmacy/services/servicesImplementation/SupplierServiceImpl.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 
 import com.pharmacy.models.Supplier;
 import com.pharmacy.services.SupplierService;
+import com.pharmacy.repositories.ProductRepository;
 import com.pharmacy.repositories.SupplierRepository;
 
 import org.springframework.stereotype.Service;
@@ -14,9 +15,32 @@ public class SupplierServiceImpl implements SupplierService {
 
   private SupplierRepository supplierRepository;
 
+  private ProductRepository productRepository;
+
   @Autowired
   public SupplierServiceImpl(SupplierRepository supplierRepository) {
     this.supplierRepository = supplierRepository;
+  }
+
+  @Autowired
+  public void setProductRepository(ProductRepository productRepository) {
+    this.productRepository = productRepository;
+  }
+
+  // Helper method to check if the supplier's phone number is valid
+  private boolean isValidContact(String contact) {
+    // Check if the contact is null or empty
+    if (contact == null || contact.trim().isEmpty()) {
+      return false;
+    }
+
+    // Check if the contact is a valid phone number (start with 04 followed by 8
+    // digits)
+    if (!contact.trim().matches("^04\\d{8}$")) {
+      return false;
+    }
+
+    return true;
   }
 
   // Get all the suppliers
@@ -41,6 +65,11 @@ public class SupplierServiceImpl implements SupplierService {
     // If the phone number is used by another supplier
     if (supplierWithSamePhoneNumber != null) {
       throw new RuntimeException("Phone number already used by another supplier");
+    }
+
+    // Check if the contact is valid
+    if (!isValidContact(supplier.getContact())) {
+      throw new RuntimeException("Invalid phone number (it must start with 04 followed by 8 digits)");
     }
 
     supplierRepository.save(supplier);
@@ -72,6 +101,11 @@ public class SupplierServiceImpl implements SupplierService {
       throw new RuntimeException("Phone number already used by another supplier");
     }
 
+    // Check if the phone number is valid
+    if (!isValidContact(supplier.getContact())) {
+      throw new RuntimeException("Invalid phone number (it must start with 04 followed by 8 digits)");
+    }
+
     supplierRepository.save(supplier);
   }
 
@@ -83,7 +117,12 @@ public class SupplierServiceImpl implements SupplierService {
 
     // If the supplier does not exist
     if (existingSupplier == null) {
-      throw new RuntimeException("Supplier does not exist");
+      throw new RuntimeException("supplier does not exist");
+    }
+
+    // Check if the supplier has associated products
+    if (productRepository.findBySupplierId(id).size() > 0) {
+      throw new RuntimeException("supplier has associated products");
     }
 
     supplierRepository.deleteById(id);
